파일 있을 때만 삭제

+++ parsing_celery/parsing/jasoseol.py
@@ -83,8 +83,9 @@ class JasoseolSailer(Sailer):
 
             recruit_store(self)
 
-        for file in self.files.values():
-            os.remove(file)
+        if self.files:
+            for file in self.files.values():
+                os.remove(file)
 
         time.sleep(random.randrange(5, 10))
